# Improve aura rendering performance