@In3luki 2024-11-10
Stealing this for #16984 . I never thought of creating a new clockwise sweep polygon to test inclusion. Very cool.