@Codas 2024-11-10
> Stealing this for #16984 .

Thank you!
I'm not sure it would be a benefit in your case though. This only improves performance for the aura renderer because the source stays the same but very many collision checks are being made to different target points. 
As evident in the 15ft aura case, the clockwise sweep polygon is actually slower when checking <15 or so target points because of the initial overhead in creating the polygon.