@In3luki 2024-11-10
I ran some tests with this PR on a desktop PC and a laptop in a quickly cobbled together scene with 1980 walls and one scene from Menace under Otari with 492 walls. 
The measured time is the time it took to complete `getAreaSquares` in `AuraRenderer#squares`.

Results with no modules (all on Chrome 130):

**PC (AMD Ryzen 5 5600X, 32 GB RAM)**

Test scene (1980 walls):
90 ft. Aura
~400ms -> ~30ms
15 ft. Aura
~5ms -> ~24ms

Beginner Box Floor 2 (492 walls):
90 ft. Aura
~100ms -> ~6ms
15 ft. Aura
~1.5ms ->  ~4ms

**Laptop (Intel i5 8265u, 16 GB RAM)**

Test scene (1980 walls):
90 ft. Aura
~700ms -> ~115ms
15 ft. Aura
~120ms -> ~90ms

Beginner Box Floor 2 (492 walls):
90 ft. Aura
~300ms -> ~35ms
15 ft. Aura
~6ms ->  ~12ms

The resulting active squares are identical in my pretty convoluted scene:

**Master**
<img width="734" alt="Master" src="https://github.com/user-attachments/assets/dc9cbe3b-8819-4402-968b-08342c88f5e4">

**PR**
<img width="733" alt="PR" src="https://github.com/user-attachments/assets/a03e0aa5-1f3e-49f3-81eb-663e04cbef74">


Looks good to me.