@CarlosFdez 2024-11-10
Thanks for doing this. A thought occured to me, and its that we might reap even more performance in the case of a token having multiple auras (a champion with marshal for example) if we pre-compute the point origin polygons using the bounds of the biggest aura on that actor, but that's probably an even more involved refactor.