@stwlam 2024-11-10
Looks great. I tested against a couple gnarly AV rooms, and the results are identical there, too. Thanks for the PR!