@Codas 2024-11-10
Thanks for the extensive additional tests all of you. I'll try to include this myself next time :)