@theripper93 2024-11-11
@Codas Funnly enough, Levels used to use quadree when it used straight walls, now it does not since canvas.edges has no quadtree, there is still the relic of the rect for the quadtree in the code waiting for a core quadtree on the edges lol https://github.com/theripper93/Levels/blob/master/scripts/handlers/sightHandler.js#L415